Move `MakeOperationGenerator` to `codegen-client`
